detecting connected module
