+++ api/opentrons/modules/__init__.py
@@ -1,8 +1,8 @@
 import os
 import logging
 import re
-from opentrons import robot, labware
-from opentrons.drivers.mag_deck import MagDeck as MagDeckDriver
+from .magdeck import MagDeck
+from opentrons import labware, robot
 
 log = logging.getLogger(__name__)
 
@@ -10,6 +10,7 @@ SUPPORTED_MODULES = {'magdeck': MagDeck}
 
 class UnsupportedModuleError(Exception):
     pass
+
 class AbsentModuleError(Exception):
     pass
 
@@ -39,8 +40,10 @@ def load(name, slot):
     else:
         raise UnsupportedModuleError("{} is not a valid module".format(name))
 
-
-def discover_modules():
+# Note: this function should be assigned to the robot.register_modules member
+# it cannot be imported and called directly inside the robot class, because
+# of the circular dependency that would create
+def discover_and_connect():
     if os.environ.get('RUNNING_ON_PI'):
         devices = os.listdir('/dev')
     else:
@@ -52,81 +55,19 @@ def discover_modules():
     for port in devices:
         match = module_port_regex.search(port)
         if match:
-            module_class = SUPPORTED_MODULES.get(match.group())
+            module_class = SUPPORTED_MODULES.get(match.group().lower())
             discovered_modules.append(module_class(port=port))
 
     log.debug('Discovered modules: {}'.format(discovered_modules))
-    return discover_modules
+    for module in discovered_modules:
+        try:
+            module.connect()
+        except AttributeError:
+            log.exception('Failed to connect module')
 
 
-class MagDeck:
-    '''
-    Under development. API subject to change
-    '''
-    def __init__(self, lw=None, port=None):
-        self.labware = lw
-        self._port = port
-        self._engaged = False
-        self._device_info = {}
 
-    def calibrate(self):
-        '''
-        Calibration involves probing for top plate to get the plate height
-        '''
-        if self._driver and self._driver.is_connected():
-            self._driver.probe_plate()
-            # return if successful or not?
-            self._engaged = False
+    return discovered_modules
 
-    def engage(self):
-        '''
-        Move the magnet to plate top - 1 mm
-        '''
-        if self._driver and self._driver.is_connected():
-            self._driver.move(self._driver.plate_height - 1.0)
-            self._engaged = True
-
-    def disengage(self):
-        '''
-        Home the magnet
-        '''
-        if self._driver and self._driver.is_connected():
-            self._driver.home()
-            self._engaged = False
-
-    @property
-    def port(self):
-        return self._port
-
-    @property
-    def device_info(self):
-        return self._device_info
-
-    @property
-    def status(self):
-        return 'engaged' if self._engaged else 'disengaged'
-
-    ######################
-    ## Internal Methods ##
-    ######################
-
-    def connect(self):
-        '''
-        Connect the serial connection
-        '''
-        if self._port:
-            self._driver = MagDeckDriver()
-            self._driver.connect(self._port)
-            self._device_info = self._driver.get_device_info()
-        else:
-            # Sanity check:
-            # Should never happen, because connect should never be called without a port on Module
-            raise MissingDevicePortError("MagDeck couldnt connect to port {}".format(self._port))
-
-    def disconnect(self):
-        '''
-        Disconnect the serial connection
-        '''
-        self._driver.disconnect()
